ezop-platform/requirements.txt
fastapi
hypercorn
python-dotenv
sqlalchemy
psycopg2-binary
